MatrixFormat enum moved into enums.hh and BatchType enum formally removed
