+++ include/blas/enums.hh
@@ -25,11 +25,6 @@ namespace batchlas {
         LAPACKE     //Belongs to NETLIB backend
     };
 
-    enum class BatchType {
-        Single,
-        Batched
-    };
-    
     enum class Transpose {
         NoTrans,
         Trans
@@ -80,14 +75,16 @@ namespace batchlas {
         TF32
     };
 
-    enum class Format {
-        COO,        //Coordinate
-        CSR,        //Compressed Sparse Row
-        CSC,        //Compressed Sparse Column
-        SELL,       //Sliced ELLPACK
-        BSR,        //Blocked Sparse Row
-        BLOCKED_ELL //Blocked ELLPACK
+    enum class MatrixFormat {
+        Dense,
+        CSR,    // Compressed Sparse Row
+        CSC,    // Compressed Sparse Column
+        COO,    // Coordinate
+        SELL,   // Sliced ELLPACK
+        BSR,    // Blocked Sparse Row
+        BLOCKED_ELL // Blocked ELLPACK
     };
+
 
     enum class Layout {
         RowMajor,
